organized the components
