+++ src/components/product/LoadingCard.tsx
