+++ src/components/product/gallery/Gallery.tsx
@@ -1,10 +1,10 @@
 'use client'
 import React from "react";
 import Image from "next/image";
-import product from '../../../public/assets/sigiriya.jpg';
+ import product from '../../../../public/assets/sigiriya.jpg';
 import { useState, useEffect } from "react";
 import { db } from "@/lib/firebase";
-import LoadingGallery from "../../loading/LoadingGallery";
+import LoadingGallery from "./LoadingGallery";
 import { getDocs, collection } from "firebase/firestore";
 
 type gData = {
